fix api

+++ API/controller/ctrlAdmin.js
@@ -10,29 +10,13 @@ const { Account } = require("../model/Account");
 const bcrypt = require("bcrypt");
 
 
-exports.renderAddProductForm = async (req, res, next) => {
-    try {
-        let categories = await Category.Category.find();
-        res.render('product/add-product', { categories: categories });
-    } catch (error) {
-        console.log(error.message);
-        res.status(500).send("Internal Server Error");
-    }
-};
 
 exports.ThemSanPham = async (req, res, next) => {
     let smg = ''
     try {
         if (req.method == "POST") {
-            let { NameProduct, Price, Size, Date, IDCategory, Image, Amount } = req.body;
-            NameProduct = NameProduct.trim();
-            Price = Price.trim();
-            Size = Size.trim();
-            Date = Date.trim();
-            IDCategory = IDCategory.trim();
-            Image = Image.trim();
-            Amount = Amount.trim();
-            if (NameProduct == '' || Price == '' || Size == '' || Date == '' || IDCategory == "" || Image == "" || Amount == "") {
+            let { NameProduct, Price, Size, Date, IDCategory,Image,Amount } = req.body;
+            if (NameProduct == '' || Price == '' || Size == '', Date == ''||IDCategory==""||Image==""||Amount=="") {
                 smg = "Không được để trống"
                 return res.status(400).json({ smg: smg })
             }
@@ -40,46 +24,34 @@ exports.ThemSanPham = async (req, res, next) => {
                 smg = "Giá phải là số"
                 return res.status(400).json({ smg: smg })
             }
-            if (isNaN(Amount)) {
-                smg = "Số lượng phải là số"
-                return res.status(400).json({ smg: smg })
-            }
-            // Validate the date using regex
-            const dateRegex = /^\d{4}-\d{2}-\d{2}$/;
-            if (!dateRegex.test(Date)) {
-                smg = "Ngày không hợp lệ, định dạng phải là YYYY-MM-DD";
-                return res.status(400).json({ smg: smg });
-            }
+            
             let objProduct = new Product.Product;
             let objProductDetail = new ProductDetail.ProductDetail;
-
-            if (req.file && fs.existsSync(req.file.path)) {
+            if (req.file!=undefined) {
                 let file_path = './public/uploads/' + req.file.originalname;
-
-                // Kiểm tra định dạng tập tin
-                if (!req.file.mimetype.startsWith('image')) {
-                    smg = 'Ảnh không đúng định dạng';
-                    fs.unlinkSync(req.file.path); // Xóa tập tin tải lên tạm thời
-                    return res.status(400).json({ smg: smg });
+                if (fs.readFileSync(req.file.path)) {
+                    // file có tồn tại
+                    console.log(req.file);
+                    if (req.file.mimetype.indexOf('image')) {
+                        smg = 'Anh khong đúng định dạng'
+                        console.log(smg)
+                        return res.status(400).json({ smg: smg })
+                    }
+                    fs.renameSync(req.file.path, file_path);
+                    objProduct.Image = 'uploads/' + req.file.originalname;
                 }
-
-                fs.renameSync(req.file.path, file_path);
-                objProduct.Image = '/uploads/' + req.file.originalname;
-            } else {
-                smg = 'Tập tin không tồn tại';
-                return res.status(400).json({ smg: smg });
+            }else{
+                objProduct.Image = Image;
             }
-
+            
             // đưa đối tượng vào cơ sở dữ liệu
             objProduct.NameProduct = NameProduct;
             objProduct.Price = Number(Price);
-            objProduct.IDCategory = IDCategory;
-
             objProductDetail.IDProduct = objProduct.id;
+            objProduct.IDCategory = IDCategory;
             objProductDetail.Size = Size.split(",");
             objProductDetail.Date = Date;
             objProductDetail.Amount = Number(Amount);
-
             await objProduct.save();
             await objProductDetail.save();
             smg = 'Thêm thành công, id mới = ' + objProduct._id
@@ -89,58 +61,32 @@ exports.ThemSanPham = async (req, res, next) => {
         console.log(error.message);
         smg = error.message;
     }
-
-    let categories = await Category.Category.find();
-    res.render('product/add-product', { smg: smg, categories: categories });
-
+    res.status(400).json({ smg: smg })
 };
 
 exports.SuaSanPham = async (req, res, next) => {
     let smg = ''
     let obj = null;
-    let objDT = null;
     try {
         obj = await Product.Product.findOne({ _id: req.params.id });
-        objDT = await ProductDetail.ProductDetail.findOne({ IDProduct: req.params.id });
-        console.log(req.query);
-        if (req.method == "POST") {
-            //console.log(req.query);
-            let { NameProduct, Price, Size, Date, IDCategory, Image, Amount } = req.body;
-            NameProduct = NameProduct.trim();
-            Price = Price.trim();
-            Size = Size.trim();
-            Date = Date.trim();
-            IDCategory = IDCategory.trim();
-            Image = Image.trim();
-            Amount = Amount.trim();
-            if (NameProduct == '' || Price == '' || IDCategory == "" || Image == "") {
+        smg = 'Lấy dữ liệu thành công'
+        if(obj==null){
+            smg = "Sản phẩm không tồn tại"
+         return res.status(400).json({  smg: smg });
+        }
+        if (req.method == "PUT") {
+            let { NameProduct, Price, Size, Date, IDCategory,Image,Amount } = req.body;
+            if (NameProduct == '' || Price == '' || Size == '', Date == ''||IDCategory==""||Image==""||Amount=="") {
                 smg = "Không được để trống"
-                return res.render('product/update-product', { smg: smg, obj: obj, objDT: objDT });
+                return res.status(400).json({ smg: smg })
             }
-            if (Size == '', Date == '' || Amount == "") {
-                return res.render('product/update-product', { smg: smg, obj: obj, objDT: objDT });
-            }
-            console.log(Price);
             if (isNaN(Price)) {
                 smg = "Giá phải là số"
-                return res.render('product/update-product', { smg: smg, obj: obj, objDT: objDT });
+                return res.status(400).json({ smg: smg })
             }
-
-            if (isNaN(Amount)) {
-                smg = "Số lượng phải là số"
-                return res.render('product/update-product', { smg: smg, obj: obj, objDT: objDT });
-            }
-
-            // Validate the date using regex
-            const dateRegex = /^\d{4}-\d{2}-\d{2}$/;
-            if (!dateRegex.test(Date)) {
-                smg = "Ngày không hợp lệ, định dạng phải là YYYY-MM-DD";
-                return res.render('product/update-product', { smg: smg, obj: obj, objDT: objDT });
-            }
-
             let objProduct = {};
             let objProductDetail = {};
-            if (req.file != undefined) {
+            if (req.file!=undefined) {
                 let file_path = './public/uploads/' + req.file.originalname;
                 if (fs.readFileSync(req.file.path)) {
                     // file có tồn tại
@@ -148,12 +94,12 @@ exports.SuaSanPham = async (req, res, next) => {
                     if (req.file.mimetype.indexOf('image')) {
                         smg = 'Anh khong đúng định dạng'
                         console.log(smg)
-                        return res.render('product/update-product', { smg: smg, obj: obj, objDT: objDT });
+                        return res.status(400).json({ smg: smg })
                     }
                     fs.renameSync(req.file.path, file_path);
                     objProduct.Image = 'uploads/' + req.file.originalname;
                 }
-            } else {
+            }else{
                 objProduct.Image = Image;
             }
             // đưa đối tượng vào cơ sở dữ liệu
@@ -164,28 +110,17 @@ exports.SuaSanPham = async (req, res, next) => {
             objProductDetail.Size = Size.split(",");
             objProductDetail.Date = Date;
             objProductDetail.Amount = Number(Amount);
-            await Product.Product.findByIdAndUpdate(req.params.id, objProduct);
-            await ProductDetail.ProductDetail.updateOne({ IDProduct: req.params.id }, objProductDetail);
+            await Product.Product.findByIdAndUpdate(req.params.id,objProduct);
+            await ProductDetail.ProductDetail.updateOne({IDProduct:req.params.id},objProductDetail);
             smg = 'Sửa thành công, id mới = ' + objProduct._id
-            //return res.status(200).json({ smg: smg })
-
-            res.render('product/update-product', { smg: smg, obj: obj, objDT: objDT });
-        }
-        else {
-            res.render('product/update-product', { smg: smg, obj: obj, objDT: objDT });
-        }
-
-
-        if (obj == null) {
-            smg = "Sản phẩm không tồn tại"
-            //return res.status(400).json({  smg: smg });
+            return res.status(200).json({ smg: smg })
         }
     } catch (error) {
         console.log(error.message);
         smg = error.message;
     }
-
-
+    res.status(400).json({ smg: smg })
+    
 };
 exports.XoaSanPham = async (req, res, next) => {
     let smg = '';
@@ -193,70 +128,56 @@ exports.XoaSanPham = async (req, res, next) => {
     let objProductDetail = null;
     try {
         objProduct = await Product.Product.findOne({ _id: req.params.id });
-        objProductDetail = await ProductDetail.ProductDetail.findOne({ IDProduct: req.params.id });
-
-        if (objProduct == null || objProductDetail == null) {
-            smg = "Sản phẩm không tồn tại";
-            return res.status(404).json({ message: smg });
+        objProductDetail = await ProductDetail.ProductDetail.findOne({IDProduct:req.params.id})
+        if(objProduct==null||objProductDetail==null){
+            smg = "Sản phẩm không tồn tại"
+         return res.status(400).json({  smg: smg });
         }
-
-        await Product.Product.findByIdAndDelete(req.params.id);
-        await ProductDetail.ProductDetail.deleteOne({ IDProduct: req.params.id });
-        smg = 'Xóa thành công';
-
-        res.render('product/delete-product');
+        smg = 'Lấy dữ liệu thành công'
+        if(req.method == 'DELETE'){
+            await Product.Product.findByIdAndDelete(req.params.id);
+            await ProductDetail.ProductDetail.deleteOne({IDProduct:req.params.id})
+            smg = 'Xóa thành công'
+        }
+       return res.status(200).json({ smg: smg});
     } catch (error) {
-        smg = "Lỗi: " + error.message;
-        return res.status(500).json({ message: smg });
+        smg = "Lỗi: "+error.message;
     }
+    res.status(400).json({ smg: smg});
 };
-
-
-exports.XemDanhSachSanPham = async (req, res, next) => {
+exports.XemDanhSachSanPham = async (req,res,next)=>{
     try {
         let list = await Product.Product.find().sort({ Name: 1 })
-        res.render('product/list-product', { listSP: list });
-    } catch (error) {
+        return res.status(200).json(list);
+       } catch (error) {
         console.log(error)
         return res.status(400).send(error)
-    }
-}
-
-exports.XemDanhSachSanPhamTheoLoai = async (req, res, next) => {
+       }
+ }
+ exports.XemDanhSachSanPhamTheoLoai =  async (req, res, next) => {
     try {
-        let list = await Product.Product.find({ CateID: req.params.id }).sort({ Name: 1 })
-        res.status(200).json(list);
+     let list = await Product.Product.find({CateID:req.params.id}).sort({ Name: 1 })
+     res.status(200).json(list);
     } catch (error) {
-        console.log(error)
-        return res.status(400).send(error)
+     console.log(error)
+     return res.status(400).send(error)
     }
-}
-exports.XemSanPham = async (req, res, next) => {
+ }
+ exports.XemSanPham = async (req, res, next) => {
     let obj = null;
     let smg = '';
     try {
         obj = await Product.Product.findOne({ _id: req.params.id });
         smg = "Lấy dữ liệu thành công"
-        if (obj == null) {
+        if(obj==null){
             smg = "Sản phẩm không tồn tại"
-            return res.status(400).json({ smg: smg });
+         return res.status(400).json({  smg: smg });
         }
     } catch (error) {
         smg = error.message;
     }
     res.status(200).json(obj);
 };
-
-exports.getCategory = async (req, res, next) => {
-    console.log("lay du lieu thanh cong")
-    try {
-        let list = await Category.Category.find()
-        res.render('category/list-category', { listCTG: list });
-    } catch (error) {
-        console.log(error)
-        return res.status(400).send(error)
-    }
-}
 
 exports.XemLoai = async (req, res, next) => {
     let obj = null;
@@ -274,27 +195,24 @@ exports.ThemLoai = async (req, res, next) => {
     let smg = ''
     try {
         if (req.method == "POST") {
-            let { NameCategory } = req.body;
-            NameCategory = NameCategory.trim();
-            if (NameCategory == '') {
+            let { NameCategory} = req.body;
+            if (NameCategory=='') {
                 smg = "Không được để trống"
-                return res.status(400).json({ smg: smg })
+                return  res.status(400).json({smg: smg })
             }
             // đưa đối tượng vào cơ sở dữ liệu
             let objCate = new Category.Category;
             objCate.NameCategory = NameCategory;
-
-            console.log("Catename= " + objCate);
-
+            console.log("Catename= "+objCate);
             await objCate.save();
             smg = 'Thêm thành công'
-            return res.status(200).json({ smg: smg })
+           return res.status(200).json({smg: smg })
         }
     } catch (error) {
         console.log(error.message);
         smg = error.message;
     }
-    res.render('category/add-category', { smg: smg });
+    res.status(400).json({smg: smg })
 };
 exports.SuaLoai = async (req, res, next) => {
     let smg = ''
@@ -302,44 +220,46 @@ exports.SuaLoai = async (req, res, next) => {
     try {
         obj = await Category.Category.findOne({ _id: req.params.id });
         smg = 'Lấy dữ liệu thành công'
-        if (obj == null) {
+        if(obj==null){
             smg = "Loại không tồn tại"
-            return res.render('category/update-category', { smg: smg, obj: obj });
+         return res.status(400).json({  smg: smg });
         }
-        if (req.method == "POST") {
-            let { NameCategory } = req.body;
-            NameCategory = NameCategory.trim();
-            if (NameCategory == '') {
+        if (req.method == "PUT") {
+            let { NameCategory} = req.body;
+            if (NameCategory=='') {
                 smg = "Không được để trống"
-                return res.status(400).json({ smg: smg })
+                return  res.status(400).json({smg: smg })
             }
             let objCate = {};
             objCate.NameCategory = NameCategory;
             await Category.Category.findByIdAndUpdate(req.params.id, objCate);
             smg = 'Sửa thành công'
         }
-        // return res.status(200).json({ smg: smg, obj: obj });
+       return res.status(200).json({ smg: smg, obj: obj });
 
 
     } catch (error) {
         smg = error.message;
 
     }
-    return res.render('category/update-category', { smg: smg, obj: obj });
+   return res.status(400).json({ smg: smg});
 };
 exports.Xoaloai = async (req, res, next) => {
     let smg = '';
     try {
         obj = await Category.Category.findOne({ _id: req.params.id });
-        if (obj == null) {
+        if(obj==null){
             smg = "Loại không tồn tại"
-            return res.status(400).json({ smg: smg });
+         return res.status(400).json({  smg: smg });
         }
-
-        await Category.Category.findByIdAndDelete(req.params.id);
-        smg = 'Xóa thành công'
-        res.render('category/delete-category');
+        smg = 'Lấy dữ liệu thành công'
+        if(req.method == 'DELETE'){
+            await Category.Category.findByIdAndDelete(req.params.id);
+            smg = 'Xóa thành công'
+        }
+        return res.status(200).json({ smg: smg});
     } catch (error) {
-        smg = "Lỗi: " + error.message;
+        smg = "Lỗi: "+error.message;
     }
+   return res.status(400).json({ smg: smg});
 };
